// src/components/base/Typography/Typography.type.ts
export type TypographyWeight =
	| "thin"
	| "extralight"
	| "light"
	| "normal"
	| "medium"
	| "semibold"
	| "bold"
	| "extrabold"
	| "black";

export type TypographyAs =
	| "h1"
	| "h2"
	| "h3"
	| "h4"
	| "h5"
	| "h6"
	| "p"
	| "span"
	| "small"
	| "legend";

export type TypographyStyling =
	| "h-36"
	| "h-32"
	| "h-24"
	| "h-20"
	| "h-18"
	| "h-16"
	| "h-14"
	| "h-12"
	| "h-10"
	| "p-16"
	| "p-14"
	| "p-12"
	| "p-10";